refactor: Names 클래스의 패키지를 car 패키지로 변경
